fix(dash): Adds `suppress_callback_exceptions=True` to Dash app
Fixes:
```
Attempting to connect a callback Input item to component:
  "table-dropdown"
but no components with that id exist in the layout.

If you are assigning callbacks to components that are
generated by other callbacks (and therefore not in the
initial layout), you can suppress this exception by setting
`suppress_callback_exceptions=True`.
This ID was used in the callback(s) for Output(s):
  amora-model-code.children

  ```